Fix nodePort service backed by pods on hybrid overlay nodes
In a cluster with hybrid overlay, after the ingress traffic is DNATed
by the ovnkube node's gateway router LB rule, the dst ip of some packets
will be changed to the ip of non-onvkube managed pods. We need to
steer this traffic to the hybrid overlay vxlan tunnel through the
node switch.

With this patch, the packects will be first forwarded from the GR to
cluster router. Then based on the routing policy, the packets will be
forwarded to the hybrid overlay bridge. At the hybrid overlay bridge
stateless SNAT rules will change the src ip to the node's Distributed
Router's.

Signed-off-by: Peng Liu <pliu@redhat.com>