Refactor hybrid-overlay-node for linux nodes
Currenly, the hybrid-overly-node binary only runs on windows nodes.
The linux binary used to be for testing long ago, and is not
used any more.

This patch refactors the linux code of hybrid-overly-node. So we can
run this binary on linux nodes which are not managed by ovnkube.
This is for the SDN live migration usecase, where we will have a
cluster with SDN and OVN-K nodes running in parallel, being connected
by Hybrid Overlay. The hybrid-overly-node on linux nodes will be
responsible for:
  1. Add Hybrid Overlay DRMAC annotation to its own node object;
  2. Remove the ovnkube pod annotations from the pods running on this
  node.

Signed-off-by: Peng Liu <pliu@redhat.com>